API & LocalStorage
Connected the site to the API for the movie data. Allowed the addition of the movies to the local storage on the site. Defaulted website to pull from localJson files when there isnt anything in local storage so first time users have access to something so they can interact with the site still.